refactor: extract shared interface logic into ucs_interface_common crate
Create a new ucs_interface_common crate that centralizes duplicated code
across FFI, gRPC, and HTTP interfaces. This makes each interface a thin
transport adapter instead of reimplementing metadata extraction, auth
resolution, header conversion, and config merge logic independently.

Modules extracted:
- metadata: MetadataPayload, get_metadata_payload, parse helpers
- auth: resolve_connector_auth, generic_auth_from_metadata + tests
- config: merge_config_with_override, merge_configs
- flow: flow_marker_to_flow_name, service_type_str
- headers: HeaderSource trait unifying HashMap and HeaderMap conversion
- request: InterfaceRequestData shared request wrapper
- middleware: extract_and_merge_config for config override
- error: InterfaceError shared error type

Also removes unused ffi dependency from grpc-server.
