+++ backend/ffi/Cargo.toml
@@ -18,6 +18,7 @@ domain_types = { path = "../domain_types" }
 common_enums = { path = "../common_enums", package = "ucs_common_enums" }
 common_utils = { path = "../common_utils", package = "ucs_common_utils" }
 ucs_env = { path = "../ucs_env" }
+ucs_interface_common = { path = "../ucs_interface_common" }
 
 # Third-party dependencies
 config = "0.14.0"
