+++ backend/ffi/src/utils.rs
@@ -1,71 +1,25 @@
 use crate::errors::FfiError;
-use common_utils::consts;
 use common_utils::metadata::{HeaderMaskingConfig, MaskedMetadata};
 use std::collections::HashMap;
 use std::sync::Arc;
-use tonic::metadata::{Ascii, MetadataMap, MetadataValue};
 use ucs_env::configs::Config;
 
 /// Converts FFI headers (HashMap) to gRPC metadata with masking support.
-/// Returns an error if any required header is absent or contains an invalid value.
+/// Delegates to the shared `headers_to_masked_metadata` implementation.
 pub fn ffi_headers_to_masked_metadata(
     headers: &HashMap<String, String>,
 ) -> Result<MaskedMetadata, FfiError> {
-    let mut metadata = MetadataMap::new();
-
-    // Required headers - these must be present
-    let required_headers = [
-        consts::X_CONNECTOR_NAME,
-        consts::X_MERCHANT_ID,
-        consts::X_REQUEST_ID,
-        consts::X_TENANT_ID,
-        consts::X_AUTH,
-    ];
-
-    // Optional headers - these may or may not be present
-    let optional_headers = [
-        consts::X_REFERENCE_ID,
-        consts::X_API_KEY,
-        consts::X_API_SECRET,
-        consts::X_KEY1,
-        consts::X_KEY2,
-        consts::X_AUTH_KEY_MAP,
-        consts::X_SHADOW_MODE,
-    ];
-
-    // Process required headers - return error if missing
-    for header_name in required_headers {
-        let header_name: &str = header_name;
-        let header_value =
-            headers
-                .get(header_name)
-                .ok_or_else(|| FfiError::MissingRequiredHeader {
-                    key: header_name.to_string(),
-                })?;
-        let metadata_value = convert_to_metadata_value(header_name, header_value)?;
-        metadata.insert(header_name, metadata_value);
-    }
-
-    // Process optional headers - skip if missing
-    for header_name in optional_headers {
-        let header_name: &str = header_name;
-        if let Some(header_value) = headers.get(header_name) {
-            if let Ok(metadata_value) = convert_to_metadata_value(header_name, header_value) {
-                metadata.insert(header_name, metadata_value);
-            }
-        }
-    }
-
-    Ok(MaskedMetadata::new(
-        metadata,
+    ucs_interface_common::headers::headers_to_masked_metadata(
+        headers,
         HeaderMaskingConfig::default(),
-    ))
-}
-
-fn convert_to_metadata_value(key: &str, value: &str) -> Result<MetadataValue<Ascii>, FfiError> {
-    MetadataValue::try_from(value).map_err(|e| FfiError::InvalidHeaderValue {
-        key: key.to_string(),
-        reason: e.to_string(),
+    )
+    .map_err(|e| match e {
+        ucs_interface_common::error::InterfaceError::MissingRequiredHeader { key } => {
+            FfiError::MissingRequiredHeader { key }
+        }
+        ucs_interface_common::error::InterfaceError::InvalidHeaderValue { key, reason } => {
+            FfiError::InvalidHeaderValue { key, reason }
+        }
     })
 }
