+++ backend/grpc-server/Cargo.toml
@@ -19,7 +19,7 @@ hyperswitch_masking = { git = "https://github.com/juspay/hyperswitch", tag = "20
 injector = { git = "https://github.com/juspay/hyperswitch", tag = "2025.11.24.0" }
 tracing-kafka = { path = "../tracing-kafka", optional = true }
 ucs_env = { path = "../ucs_env" }
-ffi = { path = "../ffi", package = "ffi"}
+ucs_interface_common = { path = "../ucs_interface_common" }
 
 # Third-party dependencies
 config = "0.14.0"
