+++ backend/grpc-server/src/config_overrides.rs
@@ -1,4 +1,3 @@
-use crate::utils::merge_config_with_override;
 use http::{Request, Response};
 use std::{
     future::Future,
@@ -52,34 +51,24 @@ where
     }
 
     fn call(&mut self, mut req: Request<Body>) -> Self::Future {
-        // Extract x-config-override header first
         let config_override = req
             .headers()
             .get("x-config-override")
-            .and_then(|h| h.to_str().map(|s| s.to_owned()).ok());
+            .and_then(|h| h.to_str().ok());
 
-        // Only process config if override header is present
-        match config_override {
-            Some(override_str) => {
-                // Merge override with default
-                let new_config =
-                    match merge_config_with_override(override_str, (*self.base_config).clone()) {
-                        Ok(cfg) => cfg,
-                        Err(e) => {
-                            let err = tonic::Status::internal(format!(
-                                "Failed to merge config with override config: {e:?}"
-                            ));
-                            let fut = async move { Err(err) };
-                            return Box::pin(fut);
-                        }
-                    };
-
-                // Insert merged config into extensions
-                req.extensions_mut().insert(new_config);
+        match ucs_interface_common::middleware::extract_and_merge_config(
+            config_override,
+            &self.base_config,
+        ) {
+            Ok(cfg) => {
+                req.extensions_mut().insert(cfg);
             }
-            None => {
-                // No override header - insert base config
-                req.extensions_mut().insert(Arc::clone(&self.base_config));
+            Err(e) => {
+                let err = tonic::Status::internal(format!(
+                    "Failed to merge config with override config: {e:?}"
+                ));
+                let fut = async move { Err(err) };
+                return Box::pin(fut);
             }
         }
