+++ backend/grpc-server/src/http/utils.rs
@@ -1,83 +1,24 @@
 use axum::{
     extract::{FromRequest, Request},
-    http::{HeaderMap, HeaderValue, StatusCode},
+    http::{HeaderMap, StatusCode},
     response::{IntoResponse, Response},
     Json,
 };
-use common_utils::consts;
 use serde::de::DeserializeOwned;
 use std::sync::Arc;
-use tonic::metadata::{Ascii, MetadataMap, MetadataValue};
+use tonic::metadata::MetadataMap;
 
 use super::error::HttpError;
 use ucs_env::configs::Config;
 
-/// Converts HTTP headers to gRPC metadata
-/// Extracts relevant headers and adds them to the gRPC metadata map
+/// Converts HTTP headers to gRPC metadata.
+/// Delegates to the shared `headers_to_metadata` implementation.
 pub fn http_headers_to_grpc_metadata(
     http_headers: &HeaderMap,
 ) -> Result<MetadataMap, Box<tonic::Status>> {
-    let mut metadata = MetadataMap::new();
-
-    // Required headers - these must be present
-    let required_headers = [
-        consts::X_CONNECTOR_NAME,
-        consts::X_MERCHANT_ID,
-        consts::X_REQUEST_ID,
-        consts::X_TENANT_ID,
-        consts::X_AUTH,
-    ];
-
-    // Optional headers - these may or may not be present
-    let optional_headers = [
-        consts::X_REFERENCE_ID,
-        consts::X_API_KEY,
-        consts::X_API_SECRET,
-        consts::X_KEY1,
-        consts::X_KEY2,
-        consts::X_AUTH_KEY_MAP,
-        consts::X_SHADOW_MODE,
-    ];
-
-    // Process required headers - fail if missing
-    for header_name in required_headers {
-        let header_value = http_headers.get(header_name).ok_or_else(|| {
-            tonic::Status::invalid_argument(format!("Missing required header: {header_name}"))
-        })?;
-
-        let metadata_value = convert_header_to_metadata(header_name, header_value)?;
-        metadata.insert(header_name, metadata_value);
-    }
-
-    // Process optional headers - skip if missing
-    for header_name in optional_headers {
-        if let Some(header_value) = http_headers.get(header_name) {
-            let metadata_value = convert_header_to_metadata(header_name, header_value)?;
-            metadata.insert(header_name, metadata_value);
-        }
-    }
-
-    Ok(metadata)
-}
-
-fn convert_header_to_metadata(
-    header_name: &str,
-    header_value: &HeaderValue,
-) -> Result<MetadataValue<Ascii>, Box<tonic::Status>> {
-    header_value
-        .to_str()
-        .map_err(|e| {
-            Box::new(tonic::Status::invalid_argument(format!(
-                "Invalid header value for {header_name}: {e}"
-            )))
-        })
-        .and_then(|s| {
-            MetadataValue::try_from(s).map_err(|e| {
-                Box::new(tonic::Status::invalid_argument(format!(
-                    "Cannot convert header {header_name} to metadata: {e}"
-                )))
-            })
-        })
+    ucs_interface_common::headers::headers_to_metadata(http_headers).map_err(|e| {
+        Box::new(tonic::Status::from(e))
+    })
 }
 
 /// Transfers config from Axum Extension to gRPC request
@@ -105,7 +46,7 @@ where
             Err(rejection) => {
                 Err(HttpError {
                     status: StatusCode::BAD_REQUEST,
-                    message: rejection.to_string(), // Use default message
+                    message: rejection.to_string(),
                 }
                 .into_response())
             }
