+++ backend/grpc-server/src/request.rs
@@ -2,15 +2,17 @@ use std::sync::Arc;
 
 use common_utils::metadata::MaskedMetadata;
 
-use crate::utils::{get_metadata_payload, MetadataPayload};
-use ucs_env::{configs, error::ResultExtGrpc};
+use crate::utils::MetadataPayload;
+use ucs_env::configs;
 
 /// Structured request data with secure metadata access.
+/// This is the gRPC-specific wrapper around `InterfaceRequestData` that
+/// provides non-optional extensions for backward compatibility.
 #[derive(Debug)]
 pub struct RequestData<T> {
     pub payload: T,
     pub extracted_metadata: MetadataPayload,
-    pub masked_metadata: MaskedMetadata, // all metadata with masking config
+    pub masked_metadata: MaskedMetadata,
     pub extensions: tonic::Extensions,
 }
 
@@ -20,20 +22,18 @@ impl<T> RequestData<T> {
         request: tonic::Request<T>,
         config: Arc<configs::Config>,
     ) -> Result<Self, tonic::Status> {
-        let (metadata, extensions, payload) = request.into_parts();
-
-        // Construct MetadataPayload from raw metadata (auth resolved from headers)
-        let metadata_payload =
-            get_metadata_payload(&metadata, config.clone()).into_grpc_status()?;
-
-        // Pass tonic metadata and config to MaskedMetadata
-        let masked_metadata = MaskedMetadata::new(metadata, config.unmasked_headers.clone());
+        let interface_data =
+            ucs_interface_common::request::InterfaceRequestData::from_grpc_request(
+                request, config,
+            )?;
 
         Ok(Self {
-            payload,
-            extracted_metadata: metadata_payload,
-            masked_metadata,
-            extensions,
+            payload: interface_data.payload,
+            extracted_metadata: interface_data.extracted_metadata,
+            masked_metadata: interface_data.masked_metadata,
+            extensions: interface_data.extensions.ok_or_else(|| {
+                tonic::Status::internal("Extensions missing from gRPC request")
+            })?,
         })
     }
 }
