+++ backend/grpc-server/src/utils.rs
@@ -1,129 +1,21 @@
-use base64::{engine::general_purpose, Engine as _};
+// Re-export shared interface logic from ucs_interface_common
+pub use ucs_interface_common::auth::*;
+pub use ucs_interface_common::config::*;
+pub use ucs_interface_common::flow::*;
+pub use ucs_interface_common::metadata::*;
+
 use common_utils::{
-    config_patch::Patch,
-    consts::{
-        self, X_API_KEY, X_API_SECRET, X_AUTH, X_AUTH_KEY_MAP, X_CONNECTOR_AUTH, X_KEY1, X_KEY2,
-        X_SHADOW_MODE,
-    },
-    errors::CustomResult,
+    consts,
     events::{Event, EventStage, FlowName, MaskedSerdeValue},
     lineage::LineageIds,
 };
-use domain_types::{
-    connector_flow::{
-        Accept, Authenticate, Authorize, Capture, CreateOrder, CreateSessionToken, DefendDispute,
-        IncrementalAuthorization, MandateRevoke, PSync, PaymentMethodToken, PostAuthenticate,
-        PreAuthenticate, RSync, Refund, RepeatPayment, SdkSessionToken, SetupMandate,
-        SubmitEvidence, Void, VoidPC,
-    },
-    connector_types,
-    errors::{ApiError, ApplicationErrorResponse},
-    router_data::ConnectorSpecificAuth,
-    utils::ForeignTryFrom,
-};
-use error_stack::{Report, ResultExt};
 use http::request::Request;
 use hyperswitch_masking;
 use serde_json::Value;
-use std::{collections::HashMap, str::FromStr, sync::Arc};
-use tonic::metadata;
-use ucs_env::configs::ConfigPatch;
-use ucs_env::{configs, logger, error::ResultExtGrpc};
+use std::{collections::HashMap, sync::Arc};
+use ucs_env::{configs, error::ResultExtGrpc};
 
 use crate::request::RequestData;
-
-pub fn service_type_str(service_type: &configs::ServiceType) -> &'static str {
-    match service_type {
-        configs::ServiceType::Grpc => "grpc",
-        configs::ServiceType::Http => "http",
-    }
-}
-
-// Helper function to map flow markers to flow names
-pub fn flow_marker_to_flow_name<F>() -> FlowName
-where
-    F: 'static,
-{
-    let type_id = std::any::TypeId::of::<F>();
-
-    if type_id == std::any::TypeId::of::<Authorize>() {
-        FlowName::Authorize
-    } else if type_id == std::any::TypeId::of::<PSync>() {
-        FlowName::Psync
-    } else if type_id == std::any::TypeId::of::<RSync>() {
-        FlowName::Rsync
-    } else if type_id == std::any::TypeId::of::<Void>() {
-        FlowName::Void
-    } else if type_id == std::any::TypeId::of::<VoidPC>() {
-        FlowName::VoidPostCapture
-    } else if type_id == std::any::TypeId::of::<Refund>() {
-        FlowName::Refund
-    } else if type_id == std::any::TypeId::of::<Capture>() {
-        FlowName::Capture
-    } else if type_id == std::any::TypeId::of::<SetupMandate>() {
-        FlowName::SetupMandate
-    } else if type_id == std::any::TypeId::of::<RepeatPayment>() {
-        FlowName::RepeatPayment
-    } else if type_id == std::any::TypeId::of::<CreateOrder>() {
-        FlowName::CreateOrder
-    } else if type_id == std::any::TypeId::of::<CreateSessionToken>() {
-        FlowName::CreateSessionToken
-    } else if type_id == std::any::TypeId::of::<Accept>() {
-        FlowName::AcceptDispute
-    } else if type_id == std::any::TypeId::of::<DefendDispute>() {
-        FlowName::DefendDispute
-    } else if type_id == std::any::TypeId::of::<SubmitEvidence>() {
-        FlowName::SubmitEvidence
-    } else if type_id == std::any::TypeId::of::<PaymentMethodToken>() {
-        FlowName::PaymentMethodToken
-    } else if type_id == std::any::TypeId::of::<PreAuthenticate>() {
-        FlowName::PreAuthenticate
-    } else if type_id == std::any::TypeId::of::<Authenticate>() {
-        FlowName::Authenticate
-    } else if type_id == std::any::TypeId::of::<PostAuthenticate>() {
-        FlowName::PostAuthenticate
-    } else if type_id == std::any::TypeId::of::<SdkSessionToken>() {
-        FlowName::SdkSessionToken
-    } else if type_id == std::any::TypeId::of::<IncrementalAuthorization>() {
-        FlowName::IncrementalAuthorization
-    } else if type_id == std::any::TypeId::of::<MandateRevoke>() {
-        FlowName::MandateRevoke
-    } else {
-        tracing::warn!("Unknown flow marker type: {}", std::any::type_name::<F>());
-        FlowName::Unknown
-    }
-}
-
-/// Extract lineage fields from header
-pub fn extract_lineage_fields_from_metadata(
-    metadata: &metadata::MetadataMap,
-    config: &configs::LineageConfig,
-) -> LineageIds<'static> {
-    if !config.enabled {
-        return LineageIds::empty(&config.field_prefix).to_owned();
-    }
-    metadata
-        .get(&config.header_name)
-        .and_then(|value| value.to_str().ok())
-        .map(|header_value| LineageIds::new(&config.field_prefix, header_value))
-        .transpose()
-        .inspect(|value| {
-            tracing::info!(
-                parsed_fields = ?value,
-                "Successfully parsed lineage header"
-            )
-        })
-        .inspect_err(|err| {
-            tracing::warn!(
-                error = %err,
-                "Failed to parse lineage header, continuing without lineage fields"
-            )
-        })
-        .ok()
-        .flatten()
-        .unwrap_or_else(|| LineageIds::empty(&config.field_prefix))
-        .to_owned()
-}
 
 /// Record the header's fields in request's trace
 pub fn record_fields_from_header<B: hyper::body::Body>(request: &Request<B>) -> tracing::Span {
@@ -151,404 +43,10 @@ pub fn record_fields_from_header<B: hyper::body::Body>(request: &Request<B>) -> 
     span
 }
 
-/// Struct to hold extracted metadata payload
-///
-/// SECURITY WARNING: This struct should only contain non-sensitive business metadata.
-/// For any sensitive data (API keys, tokens, credentials, etc.), always:
-/// 1. Wrap in hyperswitch_masking::Secret<T>
-/// 2. Extract via MaskedMetadata methods instead of adding here
-///
-#[derive(Clone, Debug)]
-pub struct MetadataPayload {
-    pub tenant_id: String,
-    pub request_id: String,
-    pub merchant_id: String,
-    pub connector: connector_types::ConnectorEnum,
-    pub lineage_ids: LineageIds<'static>,
-    pub connector_auth_type: ConnectorSpecificAuth,
-    pub reference_id: Option<String>,
-    pub shadow_mode: bool,
-    pub resource_id: Option<String>,
-}
-
-pub fn get_metadata_payload(
-    metadata: &metadata::MetadataMap,
-    server_config: Arc<configs::Config>,
-) -> CustomResult<MetadataPayload, ApplicationErrorResponse> {
-    let connector = connector_from_metadata(metadata)?;
-    let merchant_id = merchant_id_from_metadata(metadata)?;
-    let tenant_id = tenant_id_from_metadata(metadata)?;
-    let request_id = request_id_from_metadata(metadata)?;
-    let lineage_ids = extract_lineage_fields_from_metadata(metadata, &server_config.lineage);
-    let connector_auth_type = resolve_connector_auth(metadata, &connector)?;
-    let reference_id = reference_id_from_metadata(metadata)?;
-    let resource_id = resource_id_from_metadata(metadata)?;
-    let shadow_mode = shadow_mode_from_metadata(metadata);
-    Ok(MetadataPayload {
-        tenant_id,
-        request_id,
-        merchant_id,
-        connector,
-        lineage_ids,
-        connector_auth_type,
-        reference_id,
-        shadow_mode,
-        resource_id,
-    })
-}
-
-/// Extracts typed `ConnectorAuth` from the `X-Connector-Auth` header (JSON).
-///
-/// Returns `Ok(Some(...))` if header is present and valid,
-/// `Ok(None)` if header is absent (legitimate fallback case),
-/// `Err(...)` if header is present but malformed.
-fn extract_connector_auth_from_header(
-    metadata: &metadata::MetadataMap,
-) -> CustomResult<Option<grpc_api_types::payments::ConnectorAuth>, ApplicationErrorResponse> {
-    metadata
-        .get(X_CONNECTOR_AUTH)
-        .map(|value| {
-            value
-                .to_str()
-                .change_context(ApplicationErrorResponse::BadRequest(ApiError {
-                    sub_code: "INVALID_CONNECTOR_AUTH_HEADER".to_string(),
-                    error_identifier: 400,
-                    error_message: "X-Connector-Auth header contains non-ASCII characters"
-                        .to_string(),
-                    error_object: None,
-                }))
-                .and_then(|header_str| {
-                    serde_json::from_str(header_str).change_context(
-                        ApplicationErrorResponse::BadRequest(ApiError {
-                            sub_code: "INVALID_CONNECTOR_AUTH_JSON".to_string(),
-                            error_identifier: 400,
-                            error_message: "Failed to parse X-Connector-Auth header as JSON"
-                                .to_string(),
-                            error_object: None,
-                        }),
-                    )
-                })
-        })
-        .transpose()
-}
-
-/// Resolves connector auth by trying the typed `X-Connector-Auth` header first,
-/// then falling back to legacy `x-auth` / `x-api-key` / `x-key1` headers.
-fn resolve_connector_auth(
-    metadata: &metadata::MetadataMap,
-    connector: &connector_types::ConnectorEnum,
-) -> CustomResult<ConnectorSpecificAuth, ApplicationErrorResponse> {
-    extract_connector_auth_from_header(metadata)?.map_or_else(
-        || {
-            logger::debug!(
-                "X-Connector-Auth header not found, falling back to legacy headers for connector: {}",
-                connector
-            );
-            auth_from_metadata(metadata, connector)
-        },
-        |typed_auth| {
-            logger::debug!(
-                "Connector specific auth found in X-Connector-Auth header for connector: {}",
-                connector
-            );
-            ConnectorSpecificAuth::foreign_try_from(typed_auth).map_err(|_| {
-                Report::new(ApplicationErrorResponse::BadRequest(ApiError {
-                    sub_code: "AUTH_CONVERSION_FAILED".to_string(),
-                    error_identifier: 400,
-                    error_message: "Failed to convert auth from X-Connector-Auth header".to_string(),
-                    error_object: None,
-                }))
-            })
-        },
-    )
-}
-
-pub fn connector_from_metadata(
-    metadata: &metadata::MetadataMap,
-) -> CustomResult<connector_types::ConnectorEnum, ApplicationErrorResponse> {
-    parse_metadata(metadata, consts::X_CONNECTOR_NAME).and_then(|inner| {
-        connector_types::ConnectorEnum::from_str(inner).map_err(|e| {
-            Report::new(ApplicationErrorResponse::BadRequest(ApiError {
-                sub_code: "INVALID_CONNECTOR".to_string(),
-                error_identifier: 400,
-                error_message: format!("Invalid connector: {e}"),
-                error_object: None,
-            }))
-        })
-    })
-}
-
-pub fn merchant_id_from_metadata(
-    metadata: &metadata::MetadataMap,
-) -> CustomResult<String, ApplicationErrorResponse> {
-    parse_metadata(metadata, consts::X_MERCHANT_ID)
-        .map(|inner| inner.to_string())
-        .map_err(|e| {
-            Report::new(ApplicationErrorResponse::BadRequest(ApiError {
-                sub_code: "MISSING_MERCHANT_ID".to_string(),
-                error_identifier: 400,
-                error_message: format!("Missing merchant ID in request metadata: {e}"),
-                error_object: None,
-            }))
-        })
-}
-
-pub fn request_id_from_metadata(
-    metadata: &metadata::MetadataMap,
-) -> CustomResult<String, ApplicationErrorResponse> {
-    parse_metadata(metadata, consts::X_REQUEST_ID)
-        .map(|inner| inner.to_string())
-        .map_err(|e| {
-            Report::new(ApplicationErrorResponse::BadRequest(ApiError {
-                sub_code: "MISSING_REQUEST_ID".to_string(),
-                error_identifier: 400,
-                error_message: format!("Missing request ID in request metadata: {e}"),
-                error_object: None,
-            }))
-        })
-}
-
-pub fn tenant_id_from_metadata(
-    metadata: &metadata::MetadataMap,
-) -> CustomResult<String, ApplicationErrorResponse> {
-    parse_metadata(metadata, consts::X_TENANT_ID)
-        .map(|s| s.to_string())
-        .or_else(|_| Ok("DefaultTenantId".to_string()))
-}
-
-pub fn reference_id_from_metadata(
-    metadata: &metadata::MetadataMap,
-) -> CustomResult<Option<String>, ApplicationErrorResponse> {
-    parse_optional_metadata(metadata, consts::X_REFERENCE_ID).map(|s| s.map(|s| s.to_string()))
-}
-
-pub fn resource_id_from_metadata(
-    metadata: &metadata::MetadataMap,
-) -> CustomResult<Option<String>, ApplicationErrorResponse> {
-    parse_optional_metadata(metadata, consts::X_RESOURCE_ID).map(|s| s.map(|s| s.to_string()))
-}
-
-pub fn shadow_mode_from_metadata(metadata: &metadata::MetadataMap) -> bool {
-    parse_optional_metadata(metadata, X_SHADOW_MODE)
-        .ok()
-        .flatten()
-        .map(|value| value.to_lowercase() == "true")
-        .unwrap_or(false)
-}
-
-/// Extracts connector-specific auth from metadata headers.
-/// Uses the connector name to determine which variant to create.
-pub fn auth_from_metadata(
-    metadata: &metadata::MetadataMap,
-    connector: &connector_types::ConnectorEnum,
-) -> CustomResult<ConnectorSpecificAuth, ApplicationErrorResponse> {
-    let generic_auth = generic_auth_from_metadata(metadata)?;
-    ConnectorSpecificAuth::foreign_try_from((&generic_auth, connector)).map_err(|_| {
-        Report::new(ApplicationErrorResponse::BadRequest(ApiError {
-            sub_code: "AUTH_CONVERSION_FAILED".to_string(),
-            error_identifier: 400,
-            error_message: format!("Failed to convert legacy auth for connector: {}", connector),
-            error_object: None,
-        }))
-    })
-}
-
-/// Extracts generic auth type from metadata headers.
-/// This is the legacy format that uses key1, key2, etc.
-pub fn generic_auth_from_metadata(
-    metadata: &metadata::MetadataMap,
-) -> CustomResult<domain_types::router_data::ConnectorAuthType, ApplicationErrorResponse> {
-    use domain_types::router_data::ConnectorAuthType;
-
-    let auth = parse_metadata(metadata, X_AUTH)?;
-
-    #[allow(clippy::wildcard_in_or_patterns)]
-    match auth {
-        "header-key" => Ok(ConnectorAuthType::HeaderKey {
-            api_key: parse_metadata(metadata, X_API_KEY)?.to_string().into(),
-        }),
-        "body-key" => Ok(ConnectorAuthType::BodyKey {
-            api_key: parse_metadata(metadata, X_API_KEY)?.to_string().into(),
-            key1: parse_metadata(metadata, X_KEY1)?.to_string().into(),
-        }),
-        "signature-key" => Ok(ConnectorAuthType::SignatureKey {
-            api_key: parse_metadata(metadata, X_API_KEY)?.to_string().into(),
-            key1: parse_metadata(metadata, X_KEY1)?.to_string().into(),
-            api_secret: parse_metadata(metadata, X_API_SECRET)?.to_string().into(),
-        }),
-        "multi-auth-key" => Ok(ConnectorAuthType::MultiAuthKey {
-            api_key: parse_metadata(metadata, X_API_KEY)?.to_string().into(),
-            key1: parse_metadata(metadata, X_KEY1)?.to_string().into(),
-            key2: parse_metadata(metadata, X_KEY2)?.to_string().into(),
-            api_secret: parse_metadata(metadata, X_API_SECRET)?.to_string().into(),
-        }),
-        "no-key" => Ok(ConnectorAuthType::NoKey),
-        "temporary-auth" => Ok(ConnectorAuthType::TemporaryAuth),
-        "currency-auth-key" => {
-            let auth_key_map_str = parse_metadata(metadata, X_AUTH_KEY_MAP)?;
-            let auth_key_map: HashMap<
-                common_enums::enums::Currency,
-                common_utils::pii::SecretSerdeValue,
-            > = serde_json::from_str(auth_key_map_str).change_context(
-                ApplicationErrorResponse::BadRequest(ApiError {
-                    sub_code: "INVALID_AUTH_KEY_MAP".to_string(),
-                    error_identifier: 400,
-                    error_message: "Invalid auth-key-map format".to_string(),
-                    error_object: None,
-                }),
-            )?;
-            Ok(ConnectorAuthType::CurrencyAuthKey { auth_key_map })
-        }
-        "certificate-auth" | _ => Err(Report::new(ApplicationErrorResponse::BadRequest(
-            ApiError {
-                sub_code: "INVALID_AUTH_TYPE".to_string(),
-                error_identifier: 400,
-                error_message: format!("Invalid auth type: {auth}"),
-                error_object: None,
-            },
-        ))),
-    }
-}
-
-pub fn merge_config_with_override(
-    config_override: String,
-    config: configs::Config,
-) -> CustomResult<Arc<configs::Config>, ApplicationErrorResponse> {
-    match config_override.trim().is_empty() {
-        true => Ok(Arc::new(config)),
-        false => {
-            let mut override_patch: ConfigPatch = serde_json::from_str(config_override.trim())
-                .map_err(|e| {
-                    Report::new(ApplicationErrorResponse::BadRequest(ApiError {
-                        sub_code: "CANNOT_CONVERT_TO_JSON".into(),
-                        error_identifier: 400,
-                        error_message: format!("Cannot convert override config to JSON: {e}"),
-                        error_object: None,
-                    }))
-                })?;
-
-            if let Some(proxy_patch) = override_patch.proxy.as_mut() {
-                if let Some(cert_input) = proxy_patch
-                    .mitm_ca_cert
-                    .as_ref()
-                    .and_then(|value| value.as_ref())
-                {
-                    let cert_trimmed = cert_input.trim();
-
-                    let cert = if cert_trimmed.is_empty() {
-                        Err(Report::new(ApplicationErrorResponse::BadRequest(
-                            ApiError {
-                                sub_code: "INVALID_MITM_CA_CERT_BASE64".into(),
-                                error_identifier: 400,
-                                error_message: "proxy.mitm_ca_cert must be base64-encoded"
-                                    .to_string(),
-                                error_object: None,
-                            },
-                        )))
-                    } else {
-                        let sanitized: String = cert_trimmed.split_whitespace().collect();
-                        let decoded = general_purpose::STANDARD
-                            .decode(sanitized.as_bytes())
-                            .map_err(|e| {
-                                Report::new(ApplicationErrorResponse::BadRequest(ApiError {
-                                    sub_code: "INVALID_MITM_CA_CERT_BASE64".into(),
-                                    error_identifier: 400,
-                                    error_message: format!(
-                                        "Invalid base64 for proxy.mitm_ca_cert: {e}"
-                                    ),
-                                    error_object: None,
-                                }))
-                            })?;
-
-                        String::from_utf8(decoded).map_err(|e| {
-                            Report::new(ApplicationErrorResponse::BadRequest(ApiError {
-                                sub_code: "INVALID_MITM_CA_CERT_UTF8".into(),
-                                error_identifier: 400,
-                                error_message: format!(
-                                    "Decoded proxy.mitm_ca_cert is not valid UTF-8: {e}"
-                                ),
-                                error_object: None,
-                            }))
-                        })
-                    }?;
-
-                    proxy_patch.mitm_ca_cert = Some(Some(cert));
-                }
-            }
-
-            let mut merged_config = config;
-            merged_config.apply(override_patch);
-
-            tracing::info!("Config override applied successfully");
-
-            Ok(Arc::new(merged_config))
-        }
-    }
-}
-
-pub fn merge_configs(override_val: &Value, base_val: &Value) -> Value {
-    match (base_val, override_val) {
-        (Value::Object(base_map), Value::Object(override_map)) => {
-            let mut merged = base_map.clone();
-            for (key, override_value) in override_map {
-                let base_value = base_map.get(key).unwrap_or(&Value::Null);
-                merged.insert(key.clone(), merge_configs(override_value, base_value));
-            }
-            Value::Object(merged)
-        }
-        // override replaces base for primitive, null, or array
-        (_, override_val) => override_val.clone(),
-    }
-}
-
-fn parse_metadata<'a>(
-    metadata: &'a metadata::MetadataMap,
-    key: &str,
-) -> CustomResult<&'a str, ApplicationErrorResponse> {
-    metadata
-        .get(key)
-        .ok_or_else(|| {
-            Report::new(ApplicationErrorResponse::BadRequest(ApiError {
-                sub_code: "MISSING_METADATA".to_string(),
-                error_identifier: 400,
-                error_message: format!("Missing {key} in request metadata"),
-                error_object: None,
-            }))
-        })
-        .and_then(|value| {
-            value.to_str().map_err(|e| {
-                Report::new(ApplicationErrorResponse::BadRequest(ApiError {
-                    sub_code: "INVALID_METADATA".to_string(),
-                    error_identifier: 400,
-                    error_message: format!("Invalid {key} in request metadata: {e}"),
-                    error_object: None,
-                }))
-            })
-        })
-}
-
-fn parse_optional_metadata<'a>(
-    metadata: &'a metadata::MetadataMap,
-    key: &str,
-) -> CustomResult<Option<&'a str>, ApplicationErrorResponse> {
-    metadata
-        .get(key)
-        .map(|value| value.to_str())
-        .transpose()
-        .map_err(|e| {
-            Report::new(ApplicationErrorResponse::BadRequest(ApiError {
-                sub_code: "INVALID_METADATA".to_string(),
-                error_identifier: 400,
-                error_message: format!("Invalid {key} in request metadata: {e}"),
-                error_object: None,
-            }))
-        })
-}
-
 pub fn log_before_initialization<T>(
     request_data: &RequestData<T>,
     service_name: &str,
-) -> CustomResult<(), ApplicationErrorResponse>
+) -> common_utils::errors::CustomResult<(), domain_types::errors::ApplicationErrorResponse>
 where
     T: serde::Serialize,
 {
@@ -865,110 +363,4 @@ macro_rules! implement_connector_operation {
         result
     }
 }
-}
-
-#[cfg(test)]
-#[allow(clippy::expect_used, clippy::unwrap_used, clippy::panic)]
-mod tests {
-    use super::resolve_connector_auth;
-    use common_utils::consts;
-    use domain_types::{connector_types, router_data::ConnectorSpecificAuth};
-    use hyperswitch_masking::ExposeInterface;
-    use tonic::metadata::MetadataMap;
-
-    /// Build JSON for a Stripe ConnectorAuth header value.
-    ///
-    /// Format matches prost-generated serde: PascalCase enum variant, snake_case fields.
-    /// Note: We hand-craft JSON because `Secret<String>` serializes to `"***"`.
-    fn stripe_auth_json(api_key: &str) -> String {
-        format!(
-            r#"{{"auth_type":{{"Stripe":{{"api_key":"{}"}}}}}}"#,
-            api_key
-        )
-    }
-
-    /// Build a MetadataMap with a typed `X-Connector-Auth` JSON header for Stripe.
-    fn metadata_with_typed_auth(api_key: &str) -> MetadataMap {
-        let mut metadata = MetadataMap::new();
-        let json = stripe_auth_json(api_key);
-        metadata.insert(
-            consts::X_CONNECTOR_AUTH,
-            json.parse().expect("valid x-connector-auth header"),
-        );
-        metadata
-    }
-
-    /// Build a MetadataMap with legacy `x-auth` / `x-api-key` headers.
-    fn metadata_with_legacy_auth(api_key: &str) -> MetadataMap {
-        let mut metadata = MetadataMap::new();
-        metadata.insert(
-            consts::X_AUTH,
-            "header-key".parse().expect("valid x-auth header"),
-        );
-        metadata.insert(
-            consts::X_API_KEY,
-            api_key.parse().expect("valid x-api-key header"),
-        );
-        metadata
-    }
-
-    #[test]
-    fn connector_auth_resolves_from_typed_header() {
-        let metadata = metadata_with_typed_auth("typed-key-value");
-
-        let auth = resolve_connector_auth(&metadata, &connector_types::ConnectorEnum::Stripe)
-            .expect("typed header auth should resolve");
-
-        match auth {
-            ConnectorSpecificAuth::Stripe { api_key } => {
-                assert_eq!(api_key.expose(), "typed-key-value");
-            }
-            _ => panic!("expected stripe auth"),
-        }
-    }
-
-    #[test]
-    fn connector_auth_falls_back_to_legacy_headers() {
-        let metadata = metadata_with_legacy_auth("legacy-key-value");
-
-        let auth = resolve_connector_auth(&metadata, &connector_types::ConnectorEnum::Stripe)
-            .expect("legacy header auth should resolve");
-
-        match auth {
-            ConnectorSpecificAuth::Stripe { api_key } => {
-                assert_eq!(api_key.expose(), "legacy-key-value");
-            }
-            _ => panic!("expected stripe auth"),
-        }
-    }
-
-    #[test]
-    fn connector_auth_prefers_typed_header_over_legacy() {
-        let mut metadata = metadata_with_legacy_auth("legacy-key-value");
-        // Add typed header too — it should take priority
-        let json = stripe_auth_json("typed-key-value");
-        metadata.insert(
-            consts::X_CONNECTOR_AUTH,
-            json.parse().expect("valid x-connector-auth header"),
-        );
-
-        let auth = resolve_connector_auth(&metadata, &connector_types::ConnectorEnum::Stripe)
-            .expect("typed header should take precedence");
-
-        match auth {
-            ConnectorSpecificAuth::Stripe { api_key } => {
-                assert_eq!(api_key.expose(), "typed-key-value");
-            }
-            _ => panic!("expected stripe auth"),
-        }
-    }
-
-    #[test]
-    fn connector_auth_fails_when_no_auth_present() {
-        let metadata = MetadataMap::new();
-
-        let result = resolve_connector_auth(&metadata, &connector_types::ConnectorEnum::Stripe);
-
-        assert!(result.is_err());
-    }
 }
