+++ backend/ucs_interface_common/Cargo.toml
@@ -1,0 +1,26 @@
+[package]
+name = "ucs_interface_common"
+version = "0.1.0"
+edition = "2021"
+
+[dependencies]
+# First-party dependencies
+grpc-api-types = { path = "../grpc-api-types" }
+common_enums = { path = "../common_enums", package = "ucs_common_enums" }
+common_utils = { path = "../common_utils", package = "ucs_common_utils" }
+domain_types = { path = "../domain_types" }
+ucs_env = { path = "../ucs_env" }
+hyperswitch_masking = { git = "https://github.com/juspay/hyperswitch", tag = "2025.11.24.0", package = "masking" }
+
+# Third-party dependencies
+base64 = "0.21.2"
+error-stack = "0.4.1"
+serde = { version = "1.0.189", features = ["derive"] }
+serde_json = "1.0.107"
+tonic = "0.13.0"
+tracing = "0.1.40"
+http = "1.2.0"
+thiserror = "1.0.49"
+
+[lints]
+workspace = true
